refactor search-results and rename Search to SearchHomepage
