+++ src/pages/search-results.js
@@ -1,25 +1,14 @@
 import { graphql } from "gatsby"
 import React, { useEffect, useState } from "react"
-import Search from "../Components/Search"
+import Search from "../Components/SearchHomepage"
 
 export default function SearchResults(props) {
   const [input, setInput] = useState("")
-  const [isLoading, setLoading] = useState(true)
 
   useEffect(() => {
-    getUserInput()
+    const state = props.location.state
+    if (!!state) setInput(state.input)
   }, [])
-
-  const getUserInput = () => {
-    try {
-      setInput(props.location.state.input)
-      setLoading(false)
-    } catch (err) {
-      console.log(err)
-    }
-  }
-
-  if (isLoading) return <Search data={props.data} input={""} />
 
   return <Search data={props.data} input={input} />
 }
